arm64: mm: Remove sec_debug panic for init process
This patch removes the sec_debug panic handling for the init process, The removal of this code aims to prevent the kernel from panicking when the init process encounters a fault.
This change improves the robustness of the system by allowing the kernel to handle such faults more gracefully instead of halting.

Test: Booted the kernel on gts7lwifi and verified no panics during initialization.